Extract enums for discounts
