+++ saleor/graphql/discount/enums.py
@@ -1,0 +1,16 @@
+import graphene
+
+from ...discount import DiscountValueType, VoucherType
+
+
+class DiscountValueTypeEnum(graphene.Enum):
+    FIXED = DiscountValueType.FIXED
+    PERCENTAGE = DiscountValueType.PERCENTAGE
+
+
+class VoucherTypeEnum(graphene.Enum):
+    PRODUCT = VoucherType.PRODUCT
+    COLLECTION = VoucherType.COLLECTION
+    CATEGORY = VoucherType.CATEGORY
+    SHIPPING = VoucherType.SHIPPING
+    VALUE = VoucherType.VALUE
