+++ saleor/graphql/discount/resolvers.py
@@ -3,11 +3,7 @@ import graphene_django_optimizer as gql_optimizer
 from ...discount import models
 from ..utils import filter_by_query_param
 
-VOUCHER_SEARCH_FIELDS = (
-    'name', 'code', 'discount_value', 'categories__name', 'collections__name',
-    'products__name', 'min_amount_spent', 'countries', 'discount_value_type',
-    'apply_one_per_order')
-
+VOUCHER_SEARCH_FIELDS = ('name', 'code',)
 SALE_SEARCH_FIELDS = ('name', 'value', 'type')
